ir: Enable setting breakpoint on first type inequality
When debugging type canonicalization in
type_base::get_canonical_type_for, I more often than not want to know
why a type compares different to another.  Until now, I've been doing
that by stepping in the debugger.  I figure a much efficient way of
doing that is to be able to set a breakpoint on the first occurrence
of type inequality.

To do that, I am adding a few macros to use in the 'equals' functions
to return their value: ABG_RETURN(value), ABG_RETURN_REQUAL(l,r) and
ABG_RETURN_FALSE.  Those invoke a new function called
'notify_equality_failed' when the result of the comparison is false.
This allows to just set a debugger breakpoint on
'notify_equality_failed' to know when and why the type comparison
fails.

These macros invoke notify_equality_failed only if the
WITH_DEBUG_SELF_COMPARISON macro is defined.  Otherwise, they do what
the code was doing previously.  Said otherwise, this whole shebang is
enabled only when the code is configured with
--enable-debug-self-comparison.

This patch incurs no functional change.

	* src/abg-ir.cc (notify_equality_failed): Define new static
	function if WITH_DEBUG_SELF_COMPARISON is defined.
	(ABG_RETURN_EQUAL, ABG_RETURN_FALSE, ABG_RETURN): Define new macros.
	(try_canonical_compare): Use ABG_RETURN_EQUAL rather than just
	returning the result of a comparison.
	(equals): In all the overloads, use the new ABG_RETURN* macros,
	rather than just returning boolean values.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>