Add environment::{get_type_id_from_pointer,get_canonical_type_from_type_id}
When debugging self comparison issues, once the abixml file is read
back into memory, I often want to get the type-id of an artifact that
was read from abixml or get the canonical type of an artifact which
type-id is known.

Part of that information is indirectly present in the data member
abigail::reader::reader_context::m_pointer_type_id_map after the
.typeid file is loaded from file into memory.  The problem is that the
instance of abigail::reader::reader_context is transient as it's
destroyed quickly after the abixml file is read.  We want it to stay
alive longer.  So this patch moves that data member into
abigail::environment instead, along with its accessors.  The patch
then adds the new member functions
environment::{get_type_id_from_pointer,get_canonical_type_from_type_id}
to get the type-id of an artifact de-serialized from abixml and the
canonical type of an artifact for which we now the type-id string.

	* include/abg-ir.h (environment::{get_pointer_type_id_map,
	get_type_id_from_pointer, get_canonical_type_from_type_id}):
	Declare new member functions.
	* src/abg-ir.cc (environment::{get_pointer_type_id_map,
	get_type_id_from_pointer, get_canonical_type_from_type_id}):
	Define member functions.
	(environment::priv::pointer_type_id_map_): Move
	this data member here from ...
	* src/abg-reader.cc (read_context::m_pointer_type_id_map):
	... here.
	(read_context::get_pointer_type_id_map): Remove this as it's now
	defined in environment::get_pointer_type_id_map.
	(read_context::maybe_check_abixml_canonical_type_stability):
	Adjust.
	(build_type): Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>