Bug 27236 - Don't forget to emit some referenced types
Since we arranged to only emit referenced types in translation units
where they belong, it appears that in some cases we forget to emit
some referenced types.

This is because some referenced types might belong to a translation
unit that is *already* emitted by the time we detect that a type is
referenced.

To fix this correctly, we should probably have a pass that walks the
corpus to detect referenced types, so that we have their set even
before we start emitting translation units.

But for now, the patch just detects when we are emitting the last
translation unit.  In that case all the non-emitted referenced types
are emitted.  It doesn't seem to be an issue if those don't belong to
that translation unit, compared to their original (from the DWARF)
type.

	* include/abg-writer.h (write_translation_unit): Add a new
	parameter that says if we are emitting the last TU.
	* src/abg-writer.cc (write_translation_unit::{type_is_emitted,
	decl_only_type_is_emitted}): Constify these methods.
	(write_context::has_non_emitted_referenced_types): Define new
	member function using the const methods above.
	(write_translation_unit): When emitting the last TU, emit all the
	referenced types.
	(write_corpus): Set signal when emitting the last translation
	unit.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>