ir: Fix canonical type propagation cancelling
During the canonical type propagation optimization, when the
comparison of two type sub-objects fails, we need to cancel the
(potential) propagation of the canonical type of the current type
sub-object being compared.

We were not doing that in return_comparison_result, but were expecting
it.  Oops.

Fixed thus.

This helps to fix bug https://bugzilla.redhat.com/show_bug.cgi?id=1951501.

	* src/abg-ir.cc (return_comparison_result): When the comparison of
	the current type sub-object fails, clear the potentially
	propagated canonical type and remove it from the set of types with
	non confirmed propagated canonical types.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>