RHBZ1951496 - ir: Acknowledge that "void type" is not canonicalized
In the libabigail type system, the void type is a synthetic type and
is thus not canonicalized.

We forgot to mention this "void type" to
hash_as_canonical_type_or_constant as being one of the types that are
allowed to be non canonicalized in the system.  This omission violates
an assert in that function.

The patch introduces a new is_allowed_non_canonicalized_type
subroutine that defines the types that are allowed to be non
canonicalized in the system and make it recognize "void type" as such.

hash_as_canonical_type_or_constant uses the new
is_allowed_non_canonicalized_type.

This fixes https://bugzilla.redhat.com/show_bug.cgi?id=1951496

	* src/abg-ir.cc (is_allowed_non_canonicalized_type): Define new
	static function.
	(hash_as_canonical_type_or_constant): Use it.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>