Bug 27086 - Consider all C++ virtual destructors when there are many
The complete and deleting C++ destructors have the same signature.
Because the dwarf-reader re-uses the IR of functions that have the
same signature, it can happen that one of the two destructors of a
class is missed and thus not represented in the IR.  When these
destructors are virtual, that can have an impact on class comparison,
because virtual member functions are take part in class comparison,
just like data member and unlike non-virtual member functions.

This patch fixes the build_or_get_fn_decl_if_not_suppressed to avoid
"reusing" virtual destructors, based on their signature when several
are present.  Instead an IR is built for all virtual destructors that
are seen.

	* src/abg-dwarf-reader.c (build_or_get_fn_decl_if_not_suppressed):
	Do not try to re-use a virtual destructor of a class, based on its
	signature.  Several different of these can have the same
	signature, inside a given class.
	* tests/data/test-types-stability/PR27086-libstdc++.so.6.0.26:
	Add new binary test input.
	* tests/data/Makefile.am: Add the new test input to source
	distribution.
	* tests/test-types-stability.cc (elf_paths): Add the test input
	above to this harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>