Bug 28364 - libwiretap fails self comparison
In this case, thanks to all the debugging infrastructure in place,
especially the canonical type debugging infrastructure, I was able to
notice that there was a canonicalization error on a function type when
reading the libwiretap binary as in:

    $ build/tools/abidw --debug-tc /usr/lib64/libwiretap.so.11.0.8
    structural & canonical equality different for type: function type void (wtap*)
    in compare_types_during_canonicalization at: /home/dodji/git/libabigail/PR28364/src/abg-ir.cc:13575: execution should not have reached this point!
    Abandon (core dumped)
    $

When digging deeper, I noticed that, in the DWARF reader, when
building a function type, we are associating a "textual
representation" of the function type 'void (wtap*)' to its DIE (inside
the current translation) way too early.

By too early, I mean, the association was done before the function
type was fully 'built'.  Its parameters were not 'collected', for
instance.  So that means that a 'pointer to that function type' could
be formed, with a wrong representation, during the time where the
function type wasn't fully formed.  Just moving the association to
after the type was fully constructed solved the issue.

This one was hard to spot!

Later, this uncovered the fact that we could now have (and thus
serialize) member functions of unions.  And it turned out the abixml
reader didn't expect those.  Oops.  I fixed that one as well.

	* src/abg-dwarf-reader.cc (build_function_type): Associate the DIE
	representation to the constructed type once it's fully built.
	* src/abg-reader.cc (build_function_type): Support member function of unions.
	* tests/data/Makefile.am: Add the new test input files to the
	source distribution.
	* tests/data/test-diff-pkg/wireshark/wireshark-cli-3.4.9-1.fc36.x86_64-self-check-report.txt:
	Add new test input file.
	* tests/data/test-diff-pkg/wireshark/wireshark-cli-3.4.9-1.fc36.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/wireshark/wireshark-cli-debuginfo-3.4.9-1.fc36.x86_64.rpm:
	Likewise.
	* tests/test-diff-pkg.cc (in_out_specs): Add these new test input
	files to this test harness.
	* tests/data/test-read-dwarf/PR25007-sdhci.ko.abi: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>