Add debug info package for wireshark-cli-3.4.9-1.fc36.x86_64.rpm
I forgot to add the wireshark-cli-3.4.9-1.fc36.x86_64.rpm debug info
package for the test entry that uses it in tests/test-diff-pkg.cc.
It's not necessary on the x86-64 platform, but on many others, it
seems the alternate debug info contained in that package is needed.
So I am adding it in here.

	* tests/data/test-diff-pkg/wireshark/wireshark-debuginfo-3.4.9-1.fc36.x86_64.rpm:
	Add new debug info package.
	* tests/data/Makefile.am: Add it to the source distribution.
	* tests/test-diff-pkg.cc: Use the new debug info package in the
	test harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>